Separate server and client webpack config
